refactor: move the curve point gadgets into composer::point
Pure code motion: point allocation, equality, the torsion-free check,
the group arithmetic components and variable-base scalar multiplication
move out of `composer.rs` into a `point` submodule, together with their
soundness tests.
